home_iot/requirement.txt
Django==2.0
django-cors-headers==2.4.0
django-js-asset==1.1.0
djangorestframework==3.8.2
djangorestframework-simplejwt==3.2.3
get==1.0.3
Pillow==5.2.0
post==1.0.2
public==1.0.3
PyJWT==1.6.4
pytz==2018.5
query-string==1.0.2
request==1.0.2
